+++ attributes/server.rb
@@ -18,147 +18,266 @@
 # limitations under the License.
 #
 
-# Probably driven from wrapper cookbooks, environments, or roles.
-# Keep in this namespace for backwards compat
-default['rackspace_mysql']['bind_address']               = node.attribute?('cloud') && node['cloud']['local_ipv4'] ? node['cloud']['local_ipv4'] : node['ipaddress']
-default['rackspace_mysql']['port']                       = 3306
-default['rackspace_mysql']['nice']                       = 0
-
-# actual configs start here
-default['rackspace_mysql']['auto-increment-increment']        = 1
-default['rackspace_mysql']['auto-increment-offset']           = 1
-
 default['rackspace_mysql']['allow_remote_root']               = false
 default['rackspace_mysql']['remove_anonymous_users']          = false
 default['rackspace_mysql']['remove_test_database']            = false
-default['rackspace_mysql']['root_network_acl']                = nil
-default['rackspace_mysql']['tunable']['character-set-server'] = 'utf8'
-default['rackspace_mysql']['tunable']['collation-server']     = 'utf8_general_ci'
-default['rackspace_mysql']['tunable']['lower_case_table_names']  = nil
-default['rackspace_mysql']['tunable']['back_log']             = '128'
-default['rackspace_mysql']['tunable']['key_buffer_size']           = '256M'
-default['rackspace_mysql']['tunable']['myisam_sort_buffer_size']   = '8M'
-default['rackspace_mysql']['tunable']['myisam_max_sort_file_size'] = '2147483648'
-default['rackspace_mysql']['tunable']['myisam_repair_threads']     = '1'
-default['rackspace_mysql']['tunable']['myisam-recover']            = 'BACKUP'
-default['rackspace_mysql']['tunable']['max_allowed_packet']   = '16M'
-default['rackspace_mysql']['tunable']['max_connections']      = '800'
-default['rackspace_mysql']['tunable']['max_connect_errors']   = '10'
-default['rackspace_mysql']['tunable']['concurrent_insert']    = '2'
-default['rackspace_mysql']['tunable']['connect_timeout']      = '10'
-default['rackspace_mysql']['tunable']['tmp_table_size']       = '32M'
-default['rackspace_mysql']['tunable']['max_heap_table_size']  = node['rackspace_mysql']['tunable']['tmp_table_size']
-default['rackspace_mysql']['tunable']['bulk_insert_buffer_size'] = node['rackspace_mysql']['tunable']['tmp_table_size']
-default['rackspace_mysql']['tunable']['net_read_timeout']     = '30'
-default['rackspace_mysql']['tunable']['net_write_timeout']    = '30'
-default['rackspace_mysql']['tunable']['table_cache']          = '128'
-default['rackspace_mysql']['tunable']['table_open_cache']     = node['rackspace_mysql']['tunable']['table_cache'] # table_cache is deprecated
-                                                                                              # in favor of table_open_cache
-default['rackspace_mysql']['tunable']['thread_cache_size']    = 8
-default['rackspace_mysql']['tunable']['thread_concurrency']   = 10
-default['rackspace_mysql']['tunable']['thread_stack']         = '256K'
-default['rackspace_mysql']['tunable']['sort_buffer_size']     = '2M'
-default['rackspace_mysql']['tunable']['read_buffer_size']     = '128k'
-default['rackspace_mysql']['tunable']['read_rnd_buffer_size'] = '256k'
-default['rackspace_mysql']['tunable']['join_buffer_size']     = '128k'
-default['rackspace_mysql']['tunable']['wait_timeout']         = '180'
-default['rackspace_mysql']['tunable']['open-files-limit']     = '1024'
+# default['rackspace_mysql']['root_network_acl']                = nil
 
-default['rackspace_mysql']['tunable']['sql_mode'] = nil
+#
+# WARNING *** WARNING *** WARNING
+# Use of hypens and underscores is inconsistent, and is passing through from upstream
+# The key is the configuration file option, and some use - whilst others use _
+# http://bugs.mysql.com/bug.php?id=55288
+# http://stackoverflow.com/questions/7736395/mysql-configuration-hypen-or-underscore
+# http://dev.mysql.com/doc/refman/5.5/en/mysqld-option-tables.html
+#
 
-default['rackspace_mysql']['tunable']['skip-character-set-client-handshake'] = false
-default['rackspace_mysql']['tunable']['skip-name-resolve']                   = false
+#
+# See comments in the template for the layout of the hashes
+# (Though hopefully it is self-explanitory.)
+#
 
-default['rackspace_mysql']['tunable']['slave_compressed_protocol']       = 0
+#
+# Boolean Option Flags
+#
+default['rackspace_mysql']['config']['mysqld']['skip-external-locking']['bool_flag']    = true
+default['rackspace_mysql']['config']['mysqld']['skip-name-resolve']['bool_flag']        = true
+default['rackspace_mysql']['config']['mysqld']['skip-character-set-client-handshake']['bool_flag']        = true
+default['rackspace_mysql']['config']['mysqld']['log_warnings']['bool_flag']             = true
+default['rackspace_mysql']['config']['mysqld']['log-queries-not-using-indexes']['bool_flag']   = true
+default['rackspace_mysql']['config']['mysqld']['log_bin_trust_function_creators']['bool_flag'] = true
+default['rackspace_mysql']['config']['mysqld']['skip_slave_start']['bool_flag']         = true
+default['rackspace_mysql']['config']['mysqld']['read_only']['bool_flag']                = true
+default['rackspace_mysql']['config']['mysqld']['skip-innodb']['bool_flag']              = true
+default['rackspace_mysql']['config']['mysqld']['innodb_status_file']['bool_flag']       = true
+default['rackspace_mysql']['config']['mysqld']['innodb_file_per_table']['bool_flag']    = true
+default['rackspace_mysql']['config']['mysqld']['skip-innodb-doublewrite']['bool_flag']  = true
+default['rackspace_mysql']['config']['mysqld']['skip-federated']['bool_flag']           = true
+default['rackspace_mysql']['config']['mysqld']['skip-show-database']['bool_flag']       = true
+default['rackspace_mysql']['config']['mysqld']['skip-bdb']['bool_flag']                 = true
 
-default['rackspace_mysql']['tunable']['server_id']                       = nil
-default['rackspace_mysql']['tunable']['log_bin']                         = nil
-default['rackspace_mysql']['tunable']['log_bin_trust_function_creators'] = false
+default['rackspace_mysql']['config']['mysqldump']['quick']['bool_flag']                 = true
+default['rackspace_mysql']['config']['mysqldump']['quote-names']['bool_flag']           = true
 
-default['rackspace_mysql']['tunable']['relay_log']                       = nil
-default['rackspace_mysql']['tunable']['relay_log_index']                 = nil
-default['rackspace_mysql']['tunable']['log_slave_updates']               = false
+default['rackspace_mysql']['config']['mysql']['no-auto-rehash']['bool_flag']            = true
 
-default['rackspace_mysql']['tunable']['replicate_do_db']             = nil
-default['rackspace_mysql']['tunable']['replicate_do_table']          = nil
-default['rackspace_mysql']['tunable']['replicate_ignore_db']         = nil
-default['rackspace_mysql']['tunable']['replicate_ignore_table']      = nil
-default['rackspace_mysql']['tunable']['replicate_wild_do_table']     = nil
-default['rackspace_mysql']['tunable']['replicate_wild_ignore_table'] = nil
+#
+# VALUES
+#
 
-default['rackspace_mysql']['tunable']['sync_binlog']                     = 0
-default['rackspace_mysql']['tunable']['skip_slave_start']                = false
-default['rackspace_mysql']['tunable']['read_only']                       = false
+###############################
+# [mysqld]
+###############################
 
-default['rackspace_mysql']['tunable']['log_error']                       = nil
-default['rackspace_mysql']['tunable']['log_warnings']                    = false
-default['rackspace_mysql']['tunable']['log_queries_not_using_index']     = true
-default['rackspace_mysql']['tunable']['log_bin_trust_function_creators'] = false
+default['rackspace_mysql']['config']['mysqld']['user']['value']                       = 'mysql'
+default['rackspace_mysql']['config']['mysqld']['pid-file']['value']                   = '/var/run/mysqld/mysqld.pid'
 
-default['rackspace_mysql']['tunable']['innodb_log_file_size']            = '5M'
-default['rackspace_mysql']['tunable']['innodb_buffer_pool_size']         = '128M'
-default['rackspace_mysql']['tunable']['innodb_buffer_pool_instances']    = '4'
-default['rackspace_mysql']['tunable']['innodb_additional_mem_pool_size'] = '8M'
-default['rackspace_mysql']['tunable']['innodb_data_file_path']           = 'ibdata1:10M:autoextend'
-default['rackspace_mysql']['tunable']['innodb_flush_method']             = false
-default['rackspace_mysql']['tunable']['innodb_log_buffer_size']          = '8M'
-default['rackspace_mysql']['tunable']['innodb_write_io_threads']         = '4'
-default['rackspace_mysql']['tunable']['innodb_io_capacity']              = '200'
-default['rackspace_mysql']['tunable']['innodb_file_per_table']           = true
-default['rackspace_mysql']['tunable']['innodb_lock_wait_timeout']        = '60'
+# Note that client and mysqld_safe also need socket defined
+default['rackspace_mysql']['config']['mysqld']['socket']['value']                     = '/var/lib/mysql/mysql.sock'
+default['rackspace_mysql']['config']['mysqld']['port']['value']                       = 3306
+default['rackspace_mysql']['config']['mysqld']['basedir']['value']                    = '/usr'
+default['rackspace_mysql']['config']['mysqld']['datadir']['value']                    = '/var/lib/mysql'
+default['rackspace_mysql']['config']['mysqld']['tmpdir']['value']                     = '/tmp'
+default['rackspace_mysql']['config']['mysqld']['skip-external-locking']['value']      = true
+default['rackspace_mysql']['config']['mysqld']['skip-name-resolve']['value']          = false
+
+# Charset and Collation
+default['rackspace_mysql']['config']['mysqld']['character-set-server']['value']        = 'utf8'
+default['rackspace_mysql']['config']['mysqld']['collation-server']['value']            = 'utf8_general_ci'
+
+# default['rackspace_mysql']['config']['mysqld']['lower_case_table_names']['value']      = nil
+default['rackspace_mysql']['config']['mysqld']['event_scheduler']['value']              = 0
+default['rackspace_mysql']['config']['mysqld']['skip-character-set-client-handshake']['value'] = false
+
+default['rackspace_mysql']['config']['mysqld']['bind-address']['value'] = node.attribute?('cloud') && node['cloud']['local_ipv4'] ? node['cloud']['local_ipv4'] : node['ipaddress']
+
+#
+# Fine Tuning
+#
+default['rackspace_mysql']['config']['mysqld']['key_buffer_size']['value']             = '256M'
+default['rackspace_mysql']['config']['mysqld']['max_allowed_packet']['value']          = '16M'
+default['rackspace_mysql']['config']['mysqld']['thread_stack']['value']                = '256K'
+default['rackspace_mysql']['config']['mysqld']['thread_cache_size']['value']           = 8
+default['rackspace_mysql']['config']['mysqld']['sort_buffer_size']['value']            = '2M'
+default['rackspace_mysql']['config']['mysqld']['read_buffer_size']['value']            = '128k'
+default['rackspace_mysql']['config']['mysqld']['read_rnd_buffer_size']['value']        = '256k'
+default['rackspace_mysql']['config']['mysqld']['join_buffer_size']['value']            = '128k'
+default['rackspace_mysql']['config']['mysqld']['auto-increment-increment']['value']    = 1
+default['rackspace_mysql']['config']['mysqld']['auto-increment-offset']['value']       = 1
+
+default['rackspace_mysql']['config']['mysqld']['myisam-recover']['value']              = 'BACKUP'
+default['rackspace_mysql']['config']['mysqld']['max_connections']['value']             = '800'
+default['rackspace_mysql']['config']['mysqld']['max_connect_errors']['value']          = '10'
+default['rackspace_mysql']['config']['mysqld']['concurrent_insert']['value']           = '2'
+default['rackspace_mysql']['config']['mysqld']['connect_timeout']['value']             = '10'
+default['rackspace_mysql']['config']['mysqld']['wait_timeout']['value']                = '180'
+default['rackspace_mysql']['config']['mysqld']['net_read_timeout']['value']            = '30'
+default['rackspace_mysql']['config']['mysqld']['net_write_timeout']['value']           = '30'
+default['rackspace_mysql']['config']['mysqld']['back_log']['value']                    = '128'
+
+# table_cache is deprecated in favor of table_open_cache
+default['rackspace_mysql']['config']['mysqld']['table_open_cache']['value']            = '128'
+
+default['rackspace_mysql']['config']['mysqld']['tmp_table_size']['value']              = '32M'
+default['rackspace_mysql']['config']['mysqld']['max_heap_table_size']                  = node['rackspace_mysql']['config']['mysqld']['tmp_table_size']
+default['rackspace_mysql']['config']['mysqld']['bulk_insert_buffer_size']              = node['rackspace_mysql']['config']['mysqld']['tmp_table_size']
+default['rackspace_mysql']['config']['mysqld']['open-files-limit']['value']            = '1024'
+
+# Default Table Settings
+# default['rackspace_mysql']['config']['mysqld']['sql_mode']['value']                    = nil
+
+#
+# Query Cache Configuration
+#
+default['rackspace_mysql']['config']['mysqld']['query_cache_type']['value']            = '0'
+default['rackspace_mysql']['config']['mysqld']['query_cache_limit']['value']           = '1M'
+default['rackspace_mysql']['config']['mysqld']['query_cache_size']['value']            = '16M'
+
+#
+# Logging
+#
+# default['rackspace_mysql']['config']['mysqld']['log_error']['value']                   = nil
+default['rackspace_mysql']['config']['mysqld']['log_warnings']['value']                = false
+
+default['rackspace_mysql']['config']['mysqld']['long_query_time']['value']             = 2
+
+#
+# Replication
+#
+# default['rackspace_mysql']['config']['mysqld']['server_id']['value']                   = nil
+# default['rackspace_mysql']['config']['mysqld']['log_bin']['value']
+default['rackspace_mysql']['config']['mysqld']['binlog_format']['value']                   = 'statement'
+# The following are only set if log_bin is set
+default['rackspace_mysql']['config']['mysqld']['log_slave_updates']['value']               = false
+default['rackspace_mysql']['config']['mysqld']['log_bin_trust_function_creators']['value'] = false
+
+default['rackspace_mysql']['config']['mysqld']['expire_logs_days']['value']            = 10
+default['rackspace_mysql']['config']['mysqld']['max_binlog_size']['value']             = '100M'
+default['rackspace_mysql']['config']['mysqld']['binlog_cache_size']['value']           = '32K'
+default['rackspace_mysql']['config']['mysqld']['sync_binlog']['value']                 = 0
+
+# default['rackspace_mysql']['config']['mysqld']['relay_log']['value']                   = nil
+# default['rackspace_mysql']['config']['mysqld']['relay_log_index']['value']             = nil
+# default['rackspace_mysql']['config']['mysqld']['replicate_do_db']['value']             = nil
+# default['rackspace_mysql']['config']['mysqld']['replicate_do_table']['value']          = nil
+# default['rackspace_mysql']['config']['mysqld']['replicate_ignore_db']['value']         = nil
+# default['rackspace_mysql']['config']['mysqld']['replicate_ignore_table']['value']      = nil
+# default['rackspace_mysql']['config']['mysqld']['replicate_wild_do_table']['value']     = nil
+# default['rackspace_mysql']['config']['mysqld']['replicate_wild_ignore_table']['value'] = nil
+default['rackspace_mysql']['config']['mysqld']['skip_slave_start']['value']            = false
+default['rackspace_mysql']['config']['mysqld']['read_only']['value']                   = false
+# default['rackspace_mysql']['config']['mysqld']['transaction-isolation']['value']       = nil
+default['rackspace_mysql']['config']['mysqld']['slave_compressed_protocol']['value']   = 0
+
+#
+# InnoDB
+#
+default['rackspace_mysql']['config']['mysqld']['skip-innodb']['value']                   = false
+
+# The following options are only used on MySQL >= 5.5
+default['rackspace_mysql']['config']['mysqld']['innodb_write_io_threads']['value']         = '4'
+default['rackspace_mysql']['config']['mysqld']['innodb_io_capacity']['value']              = '200'
+# innodb_read_io_threads: set in the CPU conditional below
+default['rackspace_mysql']['config']['mysqld']['innodb_buffer_pool_instances']['value']    = '4'
+
+# The following options are always set
+default['rackspace_mysql']['config']['mysqld']['innodb_log_group_home_dir']['value']       = nil
+default['rackspace_mysql']['config']['mysqld']['innodb_table_locks']['value']              = true
+default['rackspace_mysql']['config']['mysqld']['innodb_lock_wait_timeout']['value']        = '60'
+# innodb_thread_concurrency: set in the CPU conditional below
+# innodb_innodb_commit_concurrency: set in the CPU conditional below
+default['rackspace_mysql']['config']['mysqld']['innodb_support_xa']['value']               = true
+default['rackspace_mysql']['config']['mysqld']['innodb_buffer_pool_size']['value']         = '128M'
+default['rackspace_mysql']['config']['mysqld']['innodb_log_file_size']['value']            = '5M'
+default['rackspace_mysql']['config']['mysqld']['innodb_additional_mem_pool_size']['value'] = '8M'
+default['rackspace_mysql']['config']['mysqld']['innodb_data_file_path']['value']           = 'ibdata1:10M:autoextend'
+default['rackspace_mysql']['config']['mysqld']['innodb_flush_log_at_trx_commit']['value']  = '1'
+
+# The following options are optional
+default['rackspace_mysql']['config']['mysqld']['innodb_log_files_in_group']['value']       = nil
+default['rackspace_mysql']['config']['mysqld']['innodb_status_file']['value']              = false
+default['rackspace_mysql']['config']['mysqld']['innodb_file_per_table']['value']           = true
+default['rackspace_mysql']['config']['mysqld']['skip-innodb-doublewrite']['value']         = false
+default['rackspace_mysql']['config']['mysqld']['innodb_flush_method']['value']             = nil
+default['rackspace_mysql']['config']['mysqld']['innodb_log_buffer_size']['value']          = '8M'
+default['rackspace_mysql']['config']['mysqld']['innodb_adaptive_flushing']['value']        = nil
+default['rackspace_mysql']['config']['mysqld']['innodb_adaptive_flushing_method']['value'] = nil
+default['rackspace_mysql']['config']['mysqld']['innodb_adaptive_checkpoint']['value']      = nil
+default['rackspace_mysql']['config']['mysqld']['innodb_change_buffering']['value']         = nil
+default['rackspace_mysql']['config']['mysqld']['innodb_doublewrite']['value']              = nil
+default['rackspace_mysql']['config']['mysqld']['innodb_file_format']['value']              = nil
+default['rackspace_mysql']['config']['mysqld']['innodb_data_home_dir']                     = node['rackspace_mysql']['config']['mysqld']['datadir']
+
+# Conditional/dynamic arguments
 if node['cpu'].nil? || node['cpu']['total'].nil?
-  default['rackspace_mysql']['tunable']['innodb_thread_concurrency']       = '8'
-  default['rackspace_mysql']['tunable']['innodb_commit_concurrency']       = '8'
-  default['rackspace_mysql']['tunable']['innodb_read_io_threads']          = '8'
+  default['rackspace_mysql']['config']['mysqld']['innodb_thread_concurrency']['value']     = '8'
+  default['rackspace_mysql']['config']['mysqld']['innodb_commit_concurrency']['value']     = '8'
+  default['rackspace_mysql']['config']['mysqld']['innodb_read_io_threads']['value']        = '8'
 else
-  default['rackspace_mysql']['tunable']['innodb_thread_concurrency']       = (node['cpu']['total'].to_i * 2).to_s
-  default['rackspace_mysql']['tunable']['innodb_commit_concurrency']       = (node['cpu']['total'].to_i * 2).to_s
-  default['rackspace_mysql']['tunable']['innodb_read_io_threads']          = (node['cpu']['total'].to_i * 2).to_s
+  default['rackspace_mysql']['config']['mysqld']['innodb_thread_concurrency']['value']     = (node['cpu']['total'].to_i * 2).to_s
+  default['rackspace_mysql']['config']['mysqld']['innodb_commit_concurrency']['value']     = (node['cpu']['total'].to_i * 2).to_s
+  default['rackspace_mysql']['config']['mysqld']['innodb_read_io_threads']['value']        = (node['cpu']['total'].to_i * 2).to_s
 end
-default['rackspace_mysql']['tunable']['innodb_flush_log_at_trx_commit']  = '1'
-default['rackspace_mysql']['tunable']['innodb_support_xa']               = true
-default['rackspace_mysql']['tunable']['innodb_table_locks']              = true
-default['rackspace_mysql']['tunable']['skip-innodb-doublewrite']         = false
 
-default['rackspace_mysql']['tunable']['transaction-isolation'] = nil
+#
+# Federated
+#
+default['rackspace_mysql']['config']['mysqld']['skip-federated']['value']                  = false
 
-default['rackspace_mysql']['tunable']['query_cache_type']    = '0'
-default['rackspace_mysql']['tunable']['query_cache_limit']    = '1M'
-default['rackspace_mysql']['tunable']['query_cache_size']     = '16M'
-
-default['rackspace_mysql']['tunable']['long_query_time']      = 2
-default['rackspace_mysql']['tunable']['expire_logs_days']     = 10
-default['rackspace_mysql']['tunable']['max_binlog_size']      = '100M'
-default['rackspace_mysql']['tunable']['binlog_cache_size']    = '32K'
-
-default['rackspace_mysql']['tmpdir'] = ['/tmp']
-
-# default['rackspace_mysql']['log_dir'] = node['rackspace_mysql']['data_dir']
-default['rackspace_mysql']['log_files_in_group'] = false
-default['rackspace_mysql']['innodb_status_file'] = false
-default['rackspace_mysql']['tunable']['innodb_change_buffering'] = false
-default['rackspace_mysql']['tunable']['innodb_doublewrite'] = false
-default['rackspace_mysql']['tunable']['innodb_file_format'] = false
-
-unless node['platform_family'] == 'rhel' && node['platform_version'].to_i < 6
-  # older RHEL platforms don't support these options
-  default['rackspace_mysql']['tunable']['event_scheduler']  = 0
-  default['rackspace_mysql']['tunable']['binlog_format']    = 'statement' if node['rackspace_mysql']['tunable']['log_bin']
-end
+#
+# Security
+#
 
 # security options
 # @see http://www.symantec.com/connect/articles/securing-mysql-step-step
 # @see http://dev.mysql.com/doc/refman/5.7/en/server-options.html#option_mysqld_chroot
-default['rackspace_mysql']['security']['chroot']                  = nil
+default['rackspace_mysql']['config']['mysqld']['chroot']['value']                  = nil
 # @see http://dev.mysql.com/doc/refman/5.7/en/server-options.html#option_mysqld_safe-user-create
-default['rackspace_mysql']['security']['safe_user_create']        = nil
+default['rackspace_mysql']['config']['mysqld']['safe_user_create']['value']        = nil
 # @see http://dev.mysql.com/doc/refman/5.7/en/server-options.html#option_mysqld_secure-auth
-default['rackspace_mysql']['security']['secure_auth']             = nil
+default['rackspace_mysql']['config']['mysqld']['secure_auth']['value']             = nil
 # @see http://dev.mysql.com/doc/refman/5.7/en/server-options.html#option_mysqld_symbolic-links
-default['rackspace_mysql']['security']['skip_symbolic_links']     = nil
+default['rackspace_mysql']['config']['mysqld']['skip_symbolic_links']['value']     = nil
 # @see http://dev.mysql.com/doc/refman/5.7/en/server-options.html#option_mysqld_secure-file-priv
-default['rackspace_mysql']['security']['secure_file_priv']        = nil
-# @see http://dev.mysql.com/doc/refman/5.7/en/server-options.html#option_mysqld_skip-show-database
-default['rackspace_mysql']['security']['skip_show_database']      = nil
+default['rackspace_mysql']['config']['mysqld']['secure_file_priv']['value']        = nil
 # @see http://dev.mysql.com/doc/refman/5.7/en/server-system-variables.html#sysvar_local_infile
-default['rackspace_mysql']['security']['local_infile']            = nil
+default['rackspace_mysql']['config']['mysqld']['local_infile']['value']            = nil
+# @see http://dev.mysql.com/doc/refman/5.7/en/server-options.html#option_mysqld_skip-show-database
+default['rackspace_mysql']['config']['mysqld']['skip-show-database']['value']      = nil
+
+#
+# BerkeleyDB
+#
+default['rackspace_mysql']['config']['mysqld']['skip-bdb']['value']                = false
+# default['rackspace_mysql']['config']['mysqld']['old_passwords']['value']         = 0
+
+###############################
+# [mysqldump]
+###############################
+default['rackspace_mysql']['config']['mysqldump']['max_allowed_packet']                = node['rackspace_mysql']['config']['mysqld']['max_allowed_packet']
+
+###############################
+# [mysql]
+###############################
+default['rackspace_mysql']['config']['mysql']['no-auto-rehash']['value']        = false
+
+###############################
+# [myisamchk]
+###############################
+default['rackspace_mysql']['config']['myisamchk']['key_buffer']                         = node['rackspace_mysql']['config']['mysqld']['max_allowed_packet']
+default['rackspace_mysql']['config']['myisamchk']['myisam_sort_buffer_size']['value']   = '8M'
+default['rackspace_mysql']['config']['myisamchk']['myisam_max_sort_file_size']['value'] = '2147483648'
+default['rackspace_mysql']['config']['myisamchk']['myisam_repair_threads']['value']     = '1'
+default['rackspace_mysql']['config']['myisamchk']['myisam-recover']                     = node['rackspace_mysql']['config']['mysqld']['myisam-recover']
+
+###############################
+# [client]
+###############################
+default['rackspace_mysql']['config']['client']['socket'] = node['rackspace_mysql']['config']['mysqld']['socket']
+default['rackspace_mysql']['config']['client']['port']   = node['rackspace_mysql']['config']['mysqld']['port']
+
+###############################
+# [mysqld_safe]
+###############################
+default['rackspace_mysql']['config']['mysqld_safe']['socket']        = node['rackspace_mysql']['config']['mysqld']['socket']
+default['rackspace_mysql']['config']['mysqld_safe']['nice']['value'] = 0